\name{add}
\alias{add}
\title{add}
\usage{
  add(x, y)
}
\arguments{
  \item{x}{A vector of numbers}

  \item{y}{A vector of numbers}
}
\value{
  The addition of \code{x} and \code{y}
}
\description{
  Adds two numbers
}
\details{
  Adds two numbers.  these can be vectors or matrices or
  otherwise compatible objects
}
\examples{
add(1,1)
add(1:3, 4:6)
}
\author{
  Michael Piccirilli
}

