% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binx.R
\name{binx}
\alias{binx}
\title{Creates Barplot of Successes in Binomial Experiment}
\usage{
binx(iter = 1000, n = 10, p = 0.5)
}
\arguments{
\item{iter, n, p}{Data Input}

\item{binx}{Function}
}
\value{
Barplot
}
\description{
Creates Barplot of Successes in Binomial Experiment
}
\examples{
binx(n = 5, p = 0.5, iter = 10000)

}
