% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cppdist}
\alias{cppdist}
\title{Calculate the euclidean distance between 2 points}
\usage{
cppdist(x1, x2, size)
}
\arguments{
\item{x1}{A numeric vector of the coordinates of x1 point.}

\item{x2}{A numeric vector of the coordinates of x2 point.}

\item{size}{A integer of the length of x1 or/and x2.}
}
\description{
Calculate the euclidean distance between 2 points
}
