% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Données de régression}
\format{
Une base de données comportant 1000 lignes et 4 colonnes

\describe{
\item{x1}{Comporte des données sur un phénomène}
\item{x2}{Comporte des données sur un phénomène}
\item{x3}{Comporte des données sur un phénomène}
\item{y}{Une variable de comptage}

}
}
\source{
Les données sont issues d'une simulation
}
\usage{
data
}
\description{
Une base de données contenant une variable de comptage y et
une matrice de variables explicative (x1, x2, x3)
}
\keyword{datasets}
