% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_sig.R
\name{est_sig}
\alias{est_sig}
\title{This function estimates the variance-covariance matrix of Y.}
\usage{
est_sig(Y, t, t.star, index)
}
\arguments{
\item{Y}{is a list}

\item{t}{is the vector of distinct time points at which Y has been observed.}

\item{t.star}{is the list of distinct time points at which Y has been observed.}

\item{index}{is the list of indices corresponding to the positions of the time points at which the data is available}
}
\value{
sig is the variance-covariance matrix of Y.

sigi is the inverse of the variance-covariance matrix of Y.
}
\description{
This function estimates the variance-covariance matrix of Y.
}
\details{
Functional Principal Component Analysis has been performed to get the variance covariance matrix for the vector obtained from Y.
}
\author{
Mityl Biswas
}
