% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.R
\name{k_means_cluster}
\alias{k_means_cluster}
\title{K-Means Clustering}
\usage{
k_means_cluster(x, k, tolerance = 1e-5, nstart = 9)
}
\arguments{
\item{x}{A numeric matrix of data}

\item{k}{A number of clusters}

\item{tolerance}{A tolerance number that determines whether the centroids are converge, default is 1e-5}

\item{nstart}{An iteration number since k-means clustering depends on the initial centroids, default is 9}
}
\value{
The original data matrix \code{x}, a vector of the clusters, a matrix of the coordinates of \code{k} centroids
}
\description{
Algorithm for kmeans clustering of numeric matrix of data
}
\examples{
x = rbind(matrix(rnorm(100, sd = 0.3), ncol = 2), matrix(rnorm(100, mean = 1, sd = 0.3), ncol = 2))
k = 2
k_means_cluster(x, k)

}
