% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ker_Xg.R
\name{ker_Xg}
\alias{ker_Xg}
\title{This function provides the gaussian kernel corresponding to X and E}
\usage{
ker_Xg(X, E)
}
\arguments{
\item{X}{is a matrix. Each column of X corresponds to q many covariates of an individual}

\item{E}{is the matrix corresponding to the component of the kernel that does not involve X.}
}
\value{
K is the kernel matrix that would have been if data were observed at all the available time points.
}
\description{
This function provides the gaussian kernel corresponding to X and E
}
\details{
The kronecker product of the kernel for X and the kernel that does not depend on X.
}
\author{
Mityl Biswas
}
