% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kern_miss.R
\name{kern_miss}
\alias{kern_miss}
\title{This function provides the kernel for the observed points}
\usage{
kern_miss(index.star, K)
}
\arguments{
\item{index.star}{is the vector of indices for the time points at which the data was observed.}

\item{K}{is the matrix corresponding to the kernel that would have been if the data were dense.}
}
\value{
K.star is the kernel matrix
}
\description{
This function provides the kernel for the observed points
}
\details{
We obtain the kernel matrix at the required indices.
}
\author{
Mityl Biswas
}
