% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{moyenne}
\alias{moyenne}
\title{Moyenne d'une variable}
\usage{
moyenne(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a number
}
\description{
Moyenne d'une variable
}
\examples{
moyenne(iris$Sepal.Length)

}
