% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Uses a sample to create a histogram showing the sample mean and confidence intervals.}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter, x, fun, alpha, cx}{Data Input}

\item{myboot2}{Function}
}
\value{
Histogram, List of Values
}
\description{
Uses a sample to create a histogram showing the sample mean and confidence intervals.
}
