% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{Uses a given sample to find the Confidence Interval at a 95% confidence}
\usage{
myci(x)
}
\arguments{
\item{x}{vector of sample being tested}

\item{myci}{Function}
}
\value{
Confidence interval with 95% confidence
}
\description{
Uses a given sample to find the Confidence Interval at a 95% confidence
}
