% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{Creates random sample and uses it to make a histogram.}
\usage{
myclt(n, iter)
}
\arguments{
\item{n, iter}{Data Input}

\item{myclt}{Function}
}
\value{
Histogram
}
\description{
Creates random sample and uses it to make a histogram.
}
