% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression.R
\name{myEigen_QR}
\alias{myEigen_QR}
\title{PCA}
\usage{
myEigen_QR(A, numIter = 1000)
}
\arguments{
\item{matrix}{A, a square matrix}

\item{integer}{numIter is the number of iterations}
}
\value{
A list with D and V where D is a vector of eigenvalues of A and V is the matrix of eigenvectors of A
}
\description{
Perform PCA on the matrix A.
}
\examples{
myEigen_QR(A, 100)
}
\seealso{
\code{\link{nchar}} which this function wraps
}
