% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression.R
\name{myLinearModel}
\alias{myLinearModel}
\title{Linear regression}
\usage{
myLinearModel(X, Y)
}
\arguments{
\item{matrix}{X is an n x p matrix of explanatory variables}

\item{matrix}{Y is an n dimensional vector of responses}
}
\value{
The least squares solution vector
}
\description{
Perform the linear regression of Y on X.
}
\examples{
myLinearModel(X, Y)
}
\seealso{
\code{\link{nchar}} which this function wraps
}
