% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression.R
\name{myLogistic}
\alias{myLogistic}
\title{Logistic regression}
\usage{
myLogistic(X, Y)
}
\arguments{
\item{matrix}{X is an n x p matrix of explanatory variables}

\item{matrix}{Y is an n dimensional vector of responses}
}
\value{
The logistic regression solution vector
}
\description{
Perform the logistic regression of Y on X.
}
\examples{
myLogistic(X, Y)
}
\seealso{
\code{\link{nchar}} which this function wraps
}
