% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlnorm.R
\name{mymlnorm}
\alias{mymlnorm}
\title{mymlNorm}
\usage{
mymlnorm(x, mu, sig, ...)
}
\arguments{
\item{x}{a vector, the trials}

\item{mu}{an integer, the mean}

\item{sig}{an integer, the standard deviation}
}
\value{
graph which will show the MLE of population mean and standard deviation
}
\description{
mymlNorm
}
