% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Create Normal Probability Distribution}
\usage{
myncurve(a, mu, sigma)
}
\arguments{
\item{a, mu, sigma}{Data Input}

\item{myncurve}{Function}
}
\value{
Distribution
}
\description{
Create Normal Probability Distribution
}
