% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression.R
\name{myQR}
\alias{myQR}
\title{QR decomposition}
\usage{
myQR(A)
}
\arguments{
\item{matrix}{a matrix of n x m}
}
\value{
A list with Q.transpose and R where Q is an orthogonal n x n matrix and R is an upper triangular n x m matrix
}
\description{
Perform QR decomposition on the matrix A.
}
\examples{
myQR(A)
}
\seealso{
\code{\link{nchar}} which this function wraps
}
