% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package1.R
\docType{package}
\name{package1}
\alias{package1}
\title{MATH 4753 Package}
\description{
Contains functions created for the University of Oklahoma MATH 4753 course
}
\author{
U. Kristine Nguyen \email{uyennhu.k.nguyen-1@ou.edu}
}
