% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package1-package.R
\docType{data}
\name{RAW_DATA_metabolomic_of_saliva}
\alias{RAW_DATA_metabolomic_of_saliva}
\title{RAW metabolomic data of saliva.}
\format{
A data frame with 79 rows and 3 variables:
\describe{
\item{Samples}{Names of samples as factor}
\item{Groups}{Control, EG-pre, and EG-post}

}
}
\source{
DMP Article: (https://www.sciencedirect.com/science/article/pii/S2352340920315560) Accessed: November 7, 2023

Excel File: (https://data.mendeley.com/datasets/wys5xd2vfg/1) Accessed: November 7, 2023
}
\usage{
data(RAW_DATA_metabolomic_of_saliva)
}
\description{
A dataset containing information about 56 metabolites, in three sub-groups of individuals.
}
\examples{
 data(RAW_DATA_metabolomic_of_saliva)
 GDPPC <- RAW_DATA_metabolomic_of_saliva$GDPPC
}
\keyword{datasets}
