% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read data in "file_name" as a data frame}
\usage{
read_data(file_name)
}
\arguments{
\item{file_name}{name of the file that contais bike counts}
}
\value{
a data frame with bike counts
}
\description{
Read data file "file name" as data frame with external packages
}
