% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.hist.R
\name{rf.hist}
\alias{rf.hist}
\title{Create Relative Frequency Histogram}
\usage{
rf.hist(v, bins = 9, main = "Relative Frequency Histogram")
}
\arguments{
\item{v, bins, main}{Data Input}

\item{rf.hist}{Function}
}
\value{
Histogram
}
\description{
Create Relative Frequency Histogram
}
