% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinar.R
\name{spinar}
\alias{spinar}
\title{Spinar(p)}
\usage{
spinar(dat, p, upper, par)
}
\arguments{
\item{dat}{: dataset}

\item{p}{: lag}

\item{upper}{: upper value for constraints}

\item{par}{: vector with values (alpha, pmf1, ...) where alpha is the true coefficient}
}
\value{

}
\description{
Spinar(p)
}
