% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_stat.R
\name{test_stat}
\alias{test_stat}
\title{This function provides the test statistic}
\usage{
test_stat(Y.sigi, K)
}
\arguments{
\item{Y.sigi}{is a vector comprising the centered observations stacked by individuals.}

\item{K}{is the kernel matrix.}
}
\value{
T is the desired test statistic.
}
\description{
This function provides the test statistic
}
\details{
The test statistic has not been centered.
}
\author{
Mityl Biswas
}
