% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolist.R
\name{tolist}
\alias{tolist}
\title{This function provides the gaussian kernel corresponding to X and E}
\usage{
tolist(Y, m.i)
}
\arguments{
\item{Y}{is the vector that needs to be converted to a list.}

\item{m.i}{is a vector comprising the lengths of each element of the resulting list.}
}
\value{
K is the kernel matrix that would have been if data were observed at all the available time points.
}
\description{
This function provides the gaussian kernel corresponding to X and E
}
\details{
The kronecker product of the kernel for X and the kernel that does not depend on X.
}
\author{
Mityl Biswas
}
