% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trivialsum.R
\name{trivialsum}
\alias{trivialsum}
\title{TrivialSum A Trivial Function}
\usage{
trivialsum(x, y)
}
\arguments{
\item{x, y}{two values to be summed}
}
\value{
the sum of x+y
}
\description{
This will simply add a few numbers together, which we could do otherwise
}
\examples{
trivialsum(2, 3)
}
