% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{My Dataset}
\format{
A data frame with ... rows and ... variables:
\describe{
\item{x1 jusqu'à x3}{Les variables explicatives permettant de
faire la régression de poisson}
\item{y}{est la variable réponse ou la variable expliquée}
...
}
}
\source{
Where the data comes from (if applicable).
}
\usage{
data
}
\description{
A description of your dataset.
}
\keyword{datasets}
