% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package1-package.R
\docType{package}
\name{package1-package}
\alias{package1}
\alias{package1-package}
\title{package1: Package used to store a data file titled energy and emissions}
\description{
Package will display data from the file titled above.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gillec1/package1}
  \item Report bugs at \url{https://github.com/gillec1/package1/issues}
}

}
\author{
\strong{Maintainer}: Claire Gillespie \email{gillec1@mcmaster.ca}

}
\keyword{internal}
