% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package1-package.R
\docType{package}
\name{package1-package}
\alias{package1}
\alias{package1-package}
\title{package1: Package used to store a data file from the article titled "Women with temporomandibular disorders: Untargeted proton nuclear magnetic resonance spectroscopy-based metabolomics of saliva and psychological instruments dataset"}
\description{
Package will display data from the article titled above.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gillec1/package1}
  \item Report bugs at \url{https://github.com/gillec1/package1/issues}
}

}
\keyword{internal}
