\name{package1-package}
\alias{package1-package}
\alias{package1}
\docType{package}
\title{
\packageTitle{package1}
}
\description{
\packageDescription{package1}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{package1}
\packageIndices{package1}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{package1}

Maintainer: \packageMaintainer{package1}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
