% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package1.R
\docType{package}
\name{Package1}
\alias{Package1}
\alias{Package1-package}
\title{A practice package created as a homework assignment for SISDIB17-3}
\description{
The package sums two numbers using the trivialsum function
}
