% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniComp.R
\name{ptruncReg}
\alias{ptruncReg}
\title{Estimates the regression coefficients in the tau-truncated Poisson regression model}
\usage{
ptruncReg(y, X, tau, tol = 1e-08, add_intercept = TRUE)
}
\arguments{
\item{y}{The vector of dependent (or response) variable}

\item{X}{a data frame or matrix containing the explanatory variables for the regression}

\item{tau}{the right-truncation threshold}

\item{tol}{the tolerance}

\item{add_intercept}{a Boolean that specifies if an intercept should be added to the model or not}
}
\value{
\code{$coeff }   : a matrix containing the maximum likelihood estimates of the regression coefficients

\code{$var_cov}  : The matrix of variance covariance

\code{$AIC}    : the value of the Akaike criterion

\code{$BIC}      : the value of the BIC criterion

\code{$n_iter}   : The number of iterations the algorithm runs until convergence
}
\description{
\code{ptruncReg} : this function fits the tau-truncated
unicomponent Poisson regression model and estimates the
regression coefficients. Note that the tau-truncated Poisson
regression model is used to analyze count data that are truncated at a known upper limit, denoted as tau.
}
\details{
The package implements a unicomponent Poisson regression model, truncated
at a known upper limit tau. The estimation is performed by maximizing
the likelihood function. This model is particularly useful for analyzing
count data with a ceiling effect.
}
\note{
It is important to ensure that the count data and truncation values are correctly
specified. The model assumes that the data follow a tau-truncated unicomponent
Poisson distribution.
}
\examples{

 library(Ex1)
# On éxécute la commande suivante  pour faire la régression
 ptruncReg(data[,1], data[,-1], tau=8, tol=1e-8, add_intercept = TRUE)

}
