% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniComp.R
\name{ptruncReg}
\alias{ptruncReg}
\title{Estimates the regression coefficients in the tau-truncated Poisson regression model}
\usage{
ptruncReg(y, X, tau, tol = 1e-08, add_intercept = TRUE)
}
\arguments{
\item{y}{The vector of dependent (or response) variable}

\item{X}{a data frame or matrix containing the explanatory variables for the regression}

\item{tau}{the right-truncation threshold}

\item{tol}{the tolerance}

\item{add_intercept}{a Boolean that specifies if an intercept should be added to the model or not}
}
\value{
\code{$coeff }   : a matrix containing the maximum likelihood estimates of the regression coefficients

\code{$var_cov}  : The matrix of variance covariance

\code{$AIC}    : the value of the Akaike criterion

\code{$BIC}      : the value of the BIC criterion

\code{$n_iter}   : The number of iterations the algorithm runs until convergence
}
\description{
Estimates the regression coefficients in the tau-truncated Poisson regression model
}
\details{
Ici se trouve tout le detail de mon package
}
\examples{
 
# On éxécute la commande suivante  pour faire la régression
 ptruncReg(data[,1], data[,-1], tau=8, tol=1e-8, add_intercept = TRUE)

}
